refactor de repositorio, entidad y esquema DB para implementación de nuevo servicio de acceso a Productos.

+++ WPFApp1/Entidades/ProductoBase.cs
@@ -41,8 +41,8 @@ namespace WPFApp1.Entidades
                 }
             }
         }
-        private int _precio;
-        public int Precio {
+        private decimal _precio;
+        public decimal Precio {
             get => _precio;
             set
             {
